# docs/swagger.yaml
basePath: /
definitions:
...
  models.Branch:
    properties:
      aashram_area:
        type: number
      address:
        type: string
      city:
        type: string
      contact_number:
        type: string
      coordinator_name:
        type: string
      country:
        type: string
      created_by:
        type: string
      created_on:
        type: string
      daily_end_time:
        type: string
      daily_start_time:
        type: string
      district:
        type: string
      email:
        type: string
      established_on:
        type: string
      id:
        type: integer
      name:
        type: string
      open_days:
        description: e.g. "Mon-Fri"
        type: string
      pincode:
        type: string
      police_station:
        type: string
      post_office:
        type: string
      state:
        type: string
      updated_by:
        type: string
      updated_on:
        type: string
    type: object
  models.BranchInfrastructure:
    properties:
      branch:
        $ref: '#/definitions/models.Branch'
      branch_id:
        type: integer
      count:
        type: integer
      created_by:
        type: string
      created_on:
        type: string
      id:
        type: integer
      type:
        type: string
      updated_by:
        type: string
      updated_on:
        type: string
    type: object
  models.BranchMember:
    properties:
      age:
        type: integer
      branch:
        $ref: '#/definitions/models.Branch'
      branch_id:
        type: integer
      branch_role:
        type: string
      created_by:
        type: string
      created_on:
        type: string
      date_of_birth:
        type: string
      date_of_samarpan:
        type: string
      id:
        type: integer
      member_type:
        type: string
      name:
        type: string
      qualification:
        type: string
      responsibility:
        type: string
      updated_by:
        type: string
      updated_on:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: vaish@gmail.com
    name: API Support
  description: API documentation for the DJJS Event Reporting Backend.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DJJS Event Reporting API
  version: "1.0"
paths:
...
  /api/branch-infra:
    get:
      description: Fetch all branch infrastructure entries
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BranchInfrastructure'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all branch infrastructure records
      tags:
      - BranchInfrastructure
    post:
      consumes:
      - application/json
      description: Adds a new infrastructure entry for a branch
      parameters:
      - description: Infrastructure payload
        in: body
        name: infra
        required: true
        schema:
          $ref: '#/definitions/models.BranchInfrastructure'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new branch infrastructure record
      tags:
      - BranchInfrastructure
  /api/branch-infra/{id}:
    delete:
      description: Delete an infrastructure entry by ID
      parameters:
      - description: Infrastructure ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a branch infrastructure record
      tags:
      - BranchInfrastructure
    put:
      consumes:
      - application/json
      description: Update existing infrastructure entry by ID
      parameters:
      - description: Infrastructure ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated fields
        in: body
        name: infra
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a branch infrastructure record
      tags:
      - BranchInfrastructure
  /api/branch-infra/branch/{branch_id}:
    get:
      description: Fetch all infrastructure entries for a specific branch
      parameters:
      - description: Branch ID
        in: path
        name: branch_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BranchInfrastructure'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get infrastructure records for a branch
      tags:
      - BranchInfrastructure
  /api/branch-member:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BranchMember'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all branch members
      tags:
      - BranchMember
    post:
      consumes:
      - application/json
      parameters:
      - description: Branch Member payload
        in: body
        name: member
        required: true
        schema:
          $ref: '#/definitions/models.BranchMember'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new branch member
      tags:
      - BranchMember
  /api/branch-member/{id}:
    delete:
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a branch member
      tags:
      - BranchMember
    put:
      consumes:
      - application/json
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated fields
        in: body
        name: member
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a branch member
      tags:
      - BranchMember
  /api/branch-member/branch/{branch_id}:
    get:
      parameters:
      - description: Branch ID
        in: path
        name: branch_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BranchMember'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get members of a branch
      tags:
      - BranchMember
  /api/branches:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Branch'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all branches
      tags:
      - Branches
    post:
      consumes:
      - application/json
      parameters:
      - description: Branch payload
        in: body
        name: branch
        required: true
        schema:
          $ref: '#/definitions/models.Branch'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new branch
      tags:
      - Branches
  /api/branches/{id}:
    delete:
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a branch
      tags:
      - Branches
    put:
      consumes:
      - application/json
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated fields
        in: body
        name: branch
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a branch
      tags:
      - Branches
  /api/branches/search:
    get:
      description: Retrieve branches by name and/or coordinator name, or list all
        if no filters.
      parameters:
      - description: Branch Name
        in: query
        name: name
        type: string
      - description: Coordinator Name
        in: query
        name: coordinator
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Branch'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get branches by name or coordinator (or all if none provided)
      tags:
      - Branches
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"